requirements.txt
authservice-grpc @ git+https://github.com/alexseipopov/capybaras_controller_authservice_grpc.git@main
certifi==2023.7.22
charset-normalizer==3.3.2
db-service-grpc @ git+https://github.com/alexseipopov/capybaras_db_handler_grpc.git@main
flake8==6.1.0
grpcio==1.59.2
grpcio-tools==1.59.2
idna==3.4
isort==5.12.0
mccabe==0.7.0
protobuf==4.25.0
pycodestyle==2.11.1
pyflakes==3.1.0
python-dotenv==1.0.0
requests==2.31.0
school-service-grpc @ git+https://github.com/alexseipopov/capybaras_school_service_grpc.git@main
urllib3==2.0.7
